backend/requirements.txt
Flask
Flask-Cors
google-generativeai
python-dotenv
scikit-learn
python-docx
pandas
openpyxl
